# backend/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Team Manager API
  description: >-
    REST API for the Teamverwaltung sports-club management application.


    Every endpoint may return `429 Too Many Requests` (global per-IP rate
    limit, `RATE_LIMIT_RPS`) or `413 Payload Too Large` (request body cap,
    4 MB by default) — see the `TooManyRequests`/`PayloadTooLarge` reusable
    responses. These aren't repeated on every operation below to keep the
    spec readable; `/auth/login` documents its own stricter rate limit
    explicitly since it differs from the global default.
  version: "1.0.0"
  contact:
    name: Team Manager
    url: https://github.com/yoadey/team-manager
servers:
  - url: /api/v1
    description: Current server
security:
  - bearerAuth: []

# ─── Tags ────────────────────────────────────────────────────────────────────
...
paths:
...
  /teams/{teamId}/absences:
    parameters:
      - $ref: "#/components/parameters/teamId"
    get:
      tags: [absences]
      operationId: listAbsences
      summary: List absences for the team
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Absence"
                  nextCursor:
                    type: string
                    nullable: true
                    description: Cursor for the next page, or null when there are no more items.
    post:
      tags: [absences]
      operationId: createAbsence
      summary: Create absence entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAbsenceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Absence"
  /teams/{teamId}/absences/mine:
    parameters:
      - $ref: "#/components/parameters/teamId"
    get:
      tags: [absences]
      operationId: listMyAbsences
      summary: List authenticated user's own absences
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Absence"
                  nextCursor:
                    type: string
                    nullable: true
                    description: Cursor for the next page, or null when there are no more items.
  /teams/{teamId}/absences/{absenceId}:
    parameters:
      - $ref: "#/components/parameters/teamId"
      - name: absenceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [absences]
      operationId: updateAbsence
      summary: Update absence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAbsenceRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Absence"
    delete:
      tags: [absences]
      operationId: deleteAbsence
      summary: Delete absence
      responses:
        "204":
          description: No Content

  # ── News ────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    teamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
...
    cursor:
      name: cursor
      in: query
      description: Opaque keyset-pagination cursor returned as nextCursor by a prior page.
      schema:
        type: string
...
  schemas:
...
    Absence:
      type: object
      required: [id, userId, from, to, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
        memberName:
          type: string
        memberAvatarColor:
          type: string
        hasPhoto:
          type: boolean
        roleColor:
          type: string
        roleName:
          type: string
    CreateAbsenceRequest:
      type: object
      required: [userId, from, to]
      properties:
        userId:
          type: string
          format: uuid
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        reason:
          type: string
    UpdateAbsenceRequest:
      type: object
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        reason:
          type: string

    # ── News ─────────────────────────────────────────────────────────────────
...